fix: update percent formatter
