+++ frappe/public/js/frappe/form/controls/float.js
@@ -28,17 +28,3 @@ frappe.ui.form.ControlFloat = class ControlFloat extends frappe.ui.form.ControlI
 		return this.df.precision || cint(frappe.boot.sysdefaults.float_precision, null);
 	}
 };
-
-frappe.ui.form.ControlPercent = class ControlPercent extends frappe.ui.form.ControlFloat {
-	format_for_input(value) {
-		if (value === null || value === undefined || isNaN(Number(value))) {
-			return "";
-		}
-		const precision = value.toString().split(".")[1]?.length || 0;
-		return format_number(
-			value,
-			this.get_number_format(),
-			Math.min(this.get_precision(), precision)
-		);
-	}
-};
