+++ frappe/public/js/frappe/form/formatters.js
@@ -94,12 +94,15 @@ frappe.form.formatters = {
 		if (value === null) {
 			return "";
 		}
-
+		const valuePrecision = value.toString().split(".")[1]?.length || 0;
 		const precision =
 			docfield.precision ||
 			cint(frappe.boot.sysdefaults && frappe.boot.sysdefaults.float_precision) ||
 			2;
-		return frappe.form.formatters._right(format_number(value, null, precision) + "%", options);
+		return frappe.form.formatters._right(
+			format_number(value, null, Math.min(precision, valuePrecision)) + "%",
+			options
+		);
 	},
 	Rating: function (value, docfield) {
 		let rating_html = "";
